back/requirements.txt
Django>=3.2
mysqlclient
djangorestframework
django-cors-headers